+++ CHANGES.md
@@ -3,6 +3,7 @@ Markdown Writer FX Change Log
 
 ## 0.7
 
+- Highlight paragraph in Preview that contain the caret of the editor.
 - Show source positions for CommonMark in Markdown AST view.